Rename onairdjango to core
