+++ core/settings/core.py
@@ -1,5 +1,5 @@
 """
-Django settings for onairdjango project.
+Django settings for core project.
 
 Generated by 'django-admin startproject' using Django 4.0.4.
 
@@ -11,7 +11,7 @@ https://docs.djangoproject.com/en/4.0/ref/settings/
 """
 import os
 from pathlib import Path
-from onairdjango.utils import *
+from core.utils import *
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent.parent
@@ -59,7 +59,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'onairdjango.urls'
+ROOT_URLCONF = 'core.urls'
 
 TEMPLATES = [
     {
@@ -79,7 +79,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'onairdjango.wsgi.application'
+WSGI_APPLICATION = 'core.wsgi.application'
 
 
 # Database
