+++ core/urls.py
@@ -1,4 +1,4 @@
-"""onairdjango URL Configuration
+"""core URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.0/topics/http/urls/
